// front/src/features/stats/types.ts
export type SiteStatsView = {
  totalArticles: number;
  totalWords: number;
  totalVisitors: number;
  totalVisits: number;
  onlineUsers: number;
  totalComments: number;
  totalFriends: number;
  totalPhotos: number;
  totalMessages: number;
  totalMoments: number;
  totalCategories: number;
  totalTags: number;
  todayVisitors: number;
  todayPageviews: number;
  yesterdayVisitors: number;
  yesterdayPageviews: number;
  monthPageviews: number;
};